# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wedding Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weddings/{weddingId}/guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List guests
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, declined, maybe]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Guest list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Create guest
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /weddings/{weddingId}/guests/import:
    post:
      operationId: importGuests
      tags: [guests]
      summary: Import guests from CSV/JSON
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /weddings/{weddingId}/guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get guest
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update guest
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "200":
          description: Guest updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete guest
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /weddings/{weddingId}/guests/{id}/rsvp:
    patch:
      operationId: updateGuestRsvp
      tags: [guests]
      summary: Update RSVP status
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpInput"
      responses:
        "200":
          description: RSVP updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /weddings/{weddingId}/guests/{id}/send-invite:
    post:
      operationId: sendGuestInvite
      tags: [guests]
      summary: Send invite via email or WhatsApp
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInviteInput"
      responses:
        "200":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendInviteResult"
...
  /weddings/{weddingId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages from guests
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Create message
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /weddings/{weddingId}/messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete message
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /weddings/{weddingId}/message-templates:
    get:
      operationId: listMessageTemplates
      tags: [messages]
      summary: List message templates
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageTemplate"
    post:
      operationId: createMessageTemplate
      tags: [messages]
      summary: Create message template
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageTemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
  /weddings/{weddingId}/message-templates/{id}:
    patch:
      operationId: updateMessageTemplate
      tags: [messages]
      summary: Update message template
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageTemplateInput"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
    delete:
      operationId: deleteMessageTemplate
      tags: [messages]
      summary: Delete message template
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /weddings/{weddingId}/send-bulk-whatsapp:
    post:
      operationId: sendBulkWhatsapp
      tags: [messages]
      summary: Send WhatsApp message to multiple guests
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkWhatsappInput"
      responses:
        "200":
          description: Bulk send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkSendResult"
...
components:
  schemas:
...
    Guest:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        group:
          type: ["string", "null"]
        rsvpStatus:
          type: string
          enum: [pending, confirmed, declined, maybe]
        plusOne:
          type: boolean
        plusOneName:
          type: ["string", "null"]
        dietaryRestrictions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        inviteSentAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - name
        - rsvpStatus
        - plusOne
        - createdAt
    GuestInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        group:
          type: ["string", "null"]
        rsvpStatus:
          type: string
          enum: [pending, confirmed, declined, maybe]
        plusOne:
          type: boolean
        plusOneName:
          type: ["string", "null"]
        dietaryRestrictions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    GuestImportInput:
      type: object
      properties:
        guests:
          type: array
          items:
            $ref: "#/components/schemas/GuestInput"
      required:
        - guests
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        errors:
          type: integer
        messages:
          type: array
          items:
            type: string
      required:
        - imported
        - errors
        - messages
    RsvpInput:
      type: object
      properties:
        rsvpStatus:
          type: string
          enum: [pending, confirmed, declined, maybe]
        dietaryRestrictions:
          type: ["string", "null"]
        plusOneName:
          type: ["string", "null"]
      required:
        - rsvpStatus
    SendInviteInput:
      type: object
      properties:
        channel:
          type: string
          enum: [email, whatsapp]
        templateId:
          type: ["integer", "null"]
      required:
        - channel
    SendInviteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Message:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        senderName:
          type: string
        messageType:
          type: string
          enum: [text, audio, video]
        content:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - senderName
        - messageType
        - createdAt
    MessageInput:
      type: object
      properties:
        senderName:
          type: string
        messageType:
          type: string
          enum: [text, audio, video]
        content:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
      required:
        - senderName
        - messageType
    MessageTemplate:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [invite, reminder, thank_you, rsvp, custom]
        content:
          type: string
        variables:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - name
        - category
        - content
        - variables
        - createdAt
    MessageTemplateInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [invite, reminder, thank_you, rsvp, custom]
        content:
          type: string
        variables:
          type: array
          items:
            type: string
    BulkWhatsappInput:
      type: object
      properties:
        guestIds:
          type: array
          items:
            type: integer
        templateId:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
      required:
        - guestIds
    BulkSendResult:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - sent
        - failed
        - errors
...